requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==21.2.0
Pillow==10.0.0 